Create a separate AppConfig file for the integration tests.

+++ Infrastructure/Services/NamePostingService.cs
@@ -25,7 +25,7 @@ namespace Infrastructure.Services
 
         protected override async Task ExecuteAsync(CancellationToken stoppingToken)
         {
-            var tweetIntervalMs = _twitterConfig.TweetIntervalSeconds * 1000;
+            var tweetIntervalMs = (int)(_twitterConfig.TweetIntervalSeconds * 1000);
             while (!stoppingToken.IsCancellationRequested)
             {
                 if (!_nameQueue.TryDequeue(out var indexedName))
